requirements.txt
# InsightSlides 依存パッケージ
# pip install -r requirements.txt

# PowerPoint操作
python-pptx>=0.6.21

# Excel操作
openpyxl>=3.1.2

# AI処理（オプション - mock以外を使う場合）
# openai>=1.0.0
# anthropic>=0.18.0

# EXE化（オプション）
# pyinstaller>=5.13.0
